+++ src/components/StripeConnectAccountForm/StripeConnectAccountForm.js
@@ -86,10 +86,11 @@ const CreateStripeAccountFields = props => {
         createResourceLocatorString('ProfilePage', routeConfiguration, { id: uuid }, {});
       const hasMarketplaceRootURL = !!marketplaceRootURL;
       const rootUrl = hasMarketplaceRootURL ? marketplaceRootURL.replace(/\/$/, '') : null;
-      const defaultBusinessURL =
+      const profilePageURL =
         hasMarketplaceRootURL && !rootUrl.includes('localhost')
           ? `${rootUrl}${pathToProfilePage(currentUserId.uuid)}`
           : `https://test-marketplace.com${pathToProfilePage(currentUserId.uuid)}`;
+      const defaultBusinessURL = `${profilePageURL}?mode=storefront`;
       form.change('businessProfileURL', defaultBusinessURL);
     }