+++ src/containers/SearchPage/SearchPage.duck.js
@@ -1,4 +1,5 @@
-import { storableError } from '../../util/errors';
+import { createImageVariantConfig } from '../../util/sdkLoader';
+import { isErrorUserPendingApproval, isForbiddenError, storableError } from '../../util/errors';
 import { convertUnitToSubUnit, unitDivisor } from '../../util/currency';
 import {
   parseDateFromISO8601,
@@ -8,8 +9,8 @@ import {
   daysBetween,
   getStartOf,
 } from '../../util/dates';
-import { createImageVariantConfig } from '../../util/sdkLoader';
 import { constructQueryParamName, isOriginInUse, isStockInUse } from '../../util/search';
+import { isUserAuthorized } from '../../util/userHelpers';
 import { parse } from '../../util/urlHelpers';
 
 import { addMarketplaceEntities } from '../../ducks/marketplaceData.duck';
@@ -41,7 +42,12 @@ const initialState = {
   currentPageResultIds: [],
 };
 
-const resultIds = data => data.data.map(l => l.id);
+const resultIds = data => {
+  const listings = data.data;
+  return listings
+    .filter(l => !l.attributes.deleted && l.attributes.state === 'published')
+    .map(l => l.id);
+};
 
 const listingPageReducer = (state = initialState, action = {}) => {
   const { type, payload } = action;
@@ -257,8 +263,11 @@ export const searchListings = (searchParams, config) => (dispatch, getState, sdk
       return response;
     })
     .catch(e => {
-      dispatch(searchListingsError(storableError(e)));
-      throw e;
+      const error = storableError(e);
+      dispatch(searchListingsError(error));
+      if (!(isErrorUserPendingApproval(error) || isForbiddenError(error))) {
+        throw e;
+      }
     });
 };
 
@@ -268,6 +277,16 @@ export const setActiveListing = listingId => ({
 });
 
 export const loadData = (params, search, config) => (dispatch, getState, sdk) => {
+  // In private marketplace mode, this page won't fetch data if the user is unauthorized
+  const state = getState();
+  const currentUser = state.user?.currentUser;
+  const isAuthorized = currentUser && isUserAuthorized(currentUser);
+  const isPrivateMarketplace = config.accessControl.marketplace.private === true;
+  const canFetchData = !isPrivateMarketplace || (isPrivateMarketplace && isAuthorized);
+  if (!canFetchData) {
+    return Promise.resolve();
+  }
+
   const queryParams = parse(search, {
     latlng: ['origin'],
     latlngBounds: ['bounds'],
@@ -294,6 +313,8 @@ export const loadData = (params, search, config) => (dispatch, getState, sdk) =>
         'title',
         'geolocation',
         'price',
+        'deleted',
+        'state',
         'publicData.listingType',
         'publicData.transactionProcessAlias',
         'publicData.unitType',